remove test_stress.py
