+++ hail/python/test/hail/table/test_table.py
@@ -9,6 +9,7 @@ from hail.utils import new_temp_file
 from hail.utils.java import Env
 from ..helpers import *
 from test.hail.matrixtable.test_file_formats import create_all_values_datasets
+from timeit import default_timer as timer
 
 setUpModule = startTestHailContext
 tearDownModule = stopTestHailContext
@@ -755,3 +756,24 @@ class Tests(unittest.TestCase):
         assert j.row.dtype == hl.tstruct(idx=hl.tint32, row1=hl.tint32, row1_1=hl.tint32)
         assert j.globals.dtype == hl.tstruct(glob1=hl.tint32, glob1_1=hl.tint32)
         j._force_count()
+
+
+def assert_time(f, max_duration):
+    start = timer()
+    x = f()
+    end = timer()
+    assert (start - end) < max_duration
+    print(start - end)
+    return x
+
+
+def test_large_number_of_fields(tmpdir):
+    mt = hl.utils.range_table(100)
+    mt = mt.annotate(**{
+        str(k): k for k in range(1000)
+    })
+    f = tmpdir.join("foo.mt")
+    assert_time(lambda: mt.count(), 5)
+    assert_time(lambda: mt.write(str(f)), 5)
+    mt = assert_time(lambda: hl.read_table(str(f)), 5)
+    assert_time(lambda: mt.count(), 5)
